Refactor cmake
