+++ deps/src/cxx_wrap/CMakeLists.txt
@@ -1,18 +1,30 @@
+cmake_minimum_required(VERSION 2.8.12)
 project(CxxWrap)
 
-cmake_minimum_required(VERSION 2.8.12)
+# Dependencies
+# ============
 
 set(CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR})
+find_package(Julia REQUIRED)
 
-set(CMAKE_INCLUDE_CURRENT_DIR ON)
-set(CMAKE_INCLUDE_CURRENT_DIR_IN_INTERFACE ON)
 set(CMAKE_MACOSX_RPATH 1)
 
-set(CxxWrap_VERSION 0.1)
+# Compilation flags
+# =================
 
-include(GenerateExportHeader)
+include(CheckCXXCompilerFlag)
 
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
+if(CMAKE_CXX_COMPILER_ID MATCHES "Clang" OR CMAKE_CXX_COMPILER_ID MATCHES "GNU" OR CMAKE_CXX_COMPILER_ID MATCHES "Intel")
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -march=native -Wunused-parameter -Wextra -Wreorder")
+    CHECK_CXX_COMPILER_FLAG("-std=c++11" HAS_CPP11_FLAG)
+
+    if (HAS_CPP11_FLAG)
+        set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
+    else()
+        message(FATAL_ERROR "Unsupported compiler -- cxx_wrap requires C++11 support!")
+    endif()
+endif()
+
 add_definitions(-DJULIA_ENABLE_THREADING)
 
 if(WIN32 AND DEFINED ENV{MSYSTEM})
@@ -21,89 +33,93 @@ if(WIN32 AND DEFINED ENV{MSYSTEM})
   endif()
 endif()
 
-find_package(Julia REQUIRED)
+# Source files
+# ============
 
-# TODO Figure out why this doesn't work on OS X:
-#set(CMAKE_CXX_STANDARD 11)
-#set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CXXWRAP_INCLUDE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/include)
+set(CXXWRAP_SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/src)
 
-add_library(cxx_wrap SHARED
-  array.hpp
-  smart_pointers.hpp
-  c_interface.cpp
-  cxx_wrap.cpp
-  functions.hpp
-  functions.cpp
-  type_conversion.hpp
-  containers/const_array.hpp
-  containers/tuple.hpp
+set(CXXWRAP_HEADERS
+    ${CXXWRAP_INCLUDE_DIR}/cxx_wrap/array.hpp
+    ${CXXWRAP_INCLUDE_DIR}/cxx_wrap/const_array.hpp
+    ${CXXWRAP_INCLUDE_DIR}/cxx_wrap/cxx_wrap.hpp
+    ${CXXWRAP_INCLUDE_DIR}/cxx_wrap/functions.hpp
+    ${CXXWRAP_INCLUDE_DIR}/cxx_wrap/smart_pointers.hpp
+    ${CXXWRAP_INCLUDE_DIR}/cxx_wrap/tuple.hpp
+    ${CXXWRAP_INCLUDE_DIR}/cxx_wrap/type_conversion.hpp
 )
 
-MESSAGE(STATUS ${Julia_INCLUDE_DIRS})
-
-target_include_directories(cxx_wrap PUBLIC ${Julia_INCLUDE_DIRS})
-generate_export_header(cxx_wrap)
-
-target_link_libraries(cxx_wrap ${Julia_LIBRARY})
-
-set_property(TARGET cxx_wrap PROPERTY VERSION ${CxxWrap_VERSION})
-set_property(TARGET cxx_wrap PROPERTY SOVERSION 0)
-set_property(TARGET cxx_wrap PROPERTY
-  INTERFACE_cxx_wrap_MAJOR_VERSION 0)
-set_property(TARGET cxx_wrap APPEND PROPERTY
-  COMPATIBLE_INTERFACE_STRING cxx_wrap_MAJOR_VERSION
+set(CXXWRAP_SOURCES
+  ${CXXWRAP_SOURCE_DIR}/c_interface.cpp
+  ${CXXWRAP_SOURCE_DIR}/cxx_wrap.cpp
+  ${CXXWRAP_SOURCE_DIR}/functions.cpp
 )
 
-install(TARGETS cxx_wrap EXPORT CxxWrapTargets
-  LIBRARY DESTINATION lib${LIBDIR_SUFFIX}
-  ARCHIVE DESTINATION lib${LIBDIR_SUFFIX}
-  RUNTIME DESTINATION lib${LIBDIR_SUFFIX}
-  INCLUDES DESTINATION include
-)
-install(
-  FILES
-    array.hpp
-    cxx_wrap.hpp
-    functions.hpp
-    smart_pointers.hpp
-    type_conversion.hpp
-  DESTINATION
-    include
-  COMPONENT
-    Devel
-)
+include_directories(${CXXWRAP_INCLUDE_DIR}/cxx_wrap)
+include_directories(${Julia_INCLUDE_DIRS})
 
-install(
-  FILES
-    containers/const_array.hpp
-    containers/tuple.hpp
-  DESTINATION
-    include/containers
-  COMPONENT
-    Devel
-)
+# Output
+# ======
 
+set(CXXWRAP_TARGET cxx_wrap)
+add_library(${CXXWRAP_TARGET} SHARED ${CXXWRAP_SOURCES} ${CXXWRAP_HEADERS})
+target_link_libraries(${CXXWRAP_TARGET} ${Julia_LIBRARY})
+set_target_properties(${CXXWRAP_TARGET} PROPERTIES
+                      PUBLIC_HEADER "${CXXWRAP_HEADERS}"
+                      COMPILE_DEFINITIONS "CXXWRAP_EXPORTS")
+
+set(CXXWRAP_NAME ${CXXWRAP_TARGET})
+if(WIN32)
+    if(MSVC)
+        set(CXXWRAP_LINK_FILE ${CXXWRAP_NAME}${CMAKE_LINK_LIBRARY_SUFFIX})
+        set(CXXWRAP_RUNTIME_FILE ${CXXWRAP_NAME}${CMAKE_SHARED_LIBRARY_SUFFIX})
+    else()
+        set(CXXWRAP_LINK_FILE lib${CXXWRAP_NAME}.dll.a)
+        set(CXXWRAP_RUNTIME_FILE lib${CXXWRAP_NAME}.dll)
+    endif()
+else()
+    if(APPLE)
+        target_link_libraries(${CXXWRAP_TARGET} "-framework CoreFoundation")
+    endif()
+    set(CXXWRAP_LINK_FILE lib${CXXWRAP_NAME}${CMAKE_SHARED_LIBRARY_SUFFIX})
+    set(CXXWRAP_RUNTIME_FILE lib${CXXWRAP_NAME}${CMAKE_SHARED_LIBRARY_SUFFIX})
+endif()
+
+include(GenerateExportHeader)
+
+generate_export_header(${CXXWRAP_TARGET})
+
+set(CxxWrap_VERSION 0.1)
+
+set_property(TARGET ${CXXWRAP_TARGET} PROPERTY VERSION ${CxxWrap_VERSION})
+set_property(TARGET ${CXXWRAP_TARGET} PROPERTY SOVERSION 0)
+set_property(TARGET ${CXXWRAP_TARGET} PROPERTY INTERFACE_cxx_wrap_MAJOR_VERSION 0)
+set_property(TARGET ${CXXWRAP_TARGET} APPEND PROPERTY COMPATIBLE_INTERFACE_STRING cxx_wrap_MAJOR_VERSION)
+
+# Installation
+# ============
+
+include(GNUInstallDirs)
 include(CMakePackageConfigHelpers)
+
+# GNUInstallDirs "DATADIR" wrong here; CMake search path wants "share".
+set(CXXWRAP_CMAKECONFIG_INSTALL_DIR "share/cmake/${PROJECT_NAME}" CACHE STRING "install path for cxx_wrapConfig.cmake")
+
+install(TARGETS ${CXXWRAP_TARGET} EXPORT CxxWrapTargets
+        ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
+        LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+        RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
+        PUBLIC_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/cxx_wrap)
+
+configure_package_config_file(${PROJECT_NAME}Config.cmake.in
+                              "${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}Config.cmake"
+                              INSTALL_DESTINATION ${CXXWRAP_CMAKECONFIG_INSTALL_DIR})
 write_basic_package_version_file(
-  "${CMAKE_CURRENT_BINARY_DIR}/CxxWrapConfigVersion.cmake"
-  VERSION ${CxxWrap_VERSION}
+  "${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}ConfigVersion.cmake"
+  VERSION ${${PROJECT_NAME}_VERSION}
   COMPATIBILITY AnyNewerVersion
 )
-
-install(EXPORT CxxWrapTargets
-  FILE
-    CxxWrapConfig.cmake
-  NAMESPACE
-    CxxWrap::
-  DESTINATION
-    lib/cmake
-)
-
-install(
-  FILES
-    "${CMAKE_CURRENT_BINARY_DIR}/CxxWrapConfigVersion.cmake"
-  DESTINATION
-    lib/cmake
-  COMPONENT
-    Devel
+install(FILES ${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}Config.cmake
+              ${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}ConfigVersion.cmake
+  DESTINATION ${CXXWRAP_CMAKECONFIG_INSTALL_DIR}
 )
