+++ deps/src/cxx_wrap/include/cxx_wrap/type_conversion.hpp
@@ -1,5 +1,5 @@
-#ifndef TYPE_CONVERSION_HPP
-#define TYPE_CONVERSION_HPP
+#ifndef CXXWRAP_TYPE_CONVERSION_HPP
+#define CXXWRAP_TYPE_CONVERSION_HPP
 
 #include <julia.h>
 #if JULIA_VERSION_MAJOR == 0 && JULIA_VERSION_MINOR > 4
