+++ deps/src/cxx_wrap/src/functions.cpp
