+++ deps/src/examples/src/containers.cpp
@@ -1,7 +1,9 @@
-#include <array.hpp>
-#include <cxx_wrap.hpp>
-#include <containers/tuple.hpp>
-#include <containers/const_array.hpp>
+#include <tuple>
+
+#include "cxx_wrap/array.hpp"
+#include "cxx_wrap/cxx_wrap.hpp"
+#include "cxx_wrap/tuple.hpp"
+#include "cxx_wrap/const_array.hpp"
 
 const double* const_vector()
 {
