+++ deps/src/examples/src/hello.cpp
@@ -1,6 +1,6 @@
 #include <string>
 
-#include <cxx_wrap.hpp>
+#include "cxx_wrap/cxx_wrap.hpp"
 
 std::string greet()
 {
