+++ deps/src/examples/src/parametric.cpp
@@ -1,6 +1,6 @@
-#include <string>
+#include <type_traits>
 
-#include <cxx_wrap.hpp>
+#include "cxx_wrap/cxx_wrap.hpp"
 
 namespace parametric
 {
